The delay calculation moved to dispatcher
