+++ src/Transport/Dispatcher.cs
@@ -8,22 +8,25 @@
     using System.Threading;
     using System.Threading.Tasks;
     using Azure.Transports.WindowsAzureStorageQueues;
+    using DelayedDelivery;
+    using DeliveryConstraints;
     using Extensibility;
     using global::Azure.Storage.Queues;
     using Logging;
     using NServiceBus.AzureStorageQueues;
+    using Performance.TimeToBeReceived;
     using Transport;
     using Unicast.Queuing;
 
     class Dispatcher : IDispatchMessages
     {
-        public Dispatcher(QueueAddressGenerator addressGenerator, AzureStorageAddressingSettings addressing, MessageWrapperSerializer serializer, Func<UnicastTransportOperation, CancellationToken, Task<bool>> shouldSend, ISubscriptionStore subscriptionStore)
+        public Dispatcher(QueueAddressGenerator addressGenerator, AzureStorageAddressingSettings addressing, MessageWrapperSerializer serializer, Func<UnicastTransportOperation, DateTimeOffset, CancellationToken, Task> scheduleMessageDelivery, ISubscriptionStore subscriptionStore)
         {
             this.subscriptionStore = subscriptionStore;
             this.addressGenerator = addressGenerator;
             this.addressing = addressing;
             this.serializer = serializer;
-            this.shouldSend = shouldSend;
+            this.scheduleMessageDelivery = scheduleMessageDelivery;
         }
 
         public async Task Dispatch(TransportOperations outgoingMessages, TransportTransaction transaction, ContextBag context)
@@ -74,9 +77,15 @@
                 Logger.DebugFormat("Sending message (ID: '{0}') to {1}", operation.Message.MessageId, operation.Destination);
             }
 
-            var dispatchDecision = await shouldSend(operation, cancellationToken).ConfigureAwait(false);
-            if (dispatchDecision == false)
+            var delay = GetVisibilityDelay(operation.DeliveryConstraints);
+            if (delay != null)
             {
+                if (FirstOrDefault<DiscardIfNotReceivedBefore>(operation.DeliveryConstraints) != null)
+                {
+                    throw new Exception($"Postponed delivery of messages with TimeToBeReceived set is not supported. Remove the TimeToBeReceived attribute to postpone messages of type '{operation.Message.Headers[Headers.EnclosedMessageTypes]}'.");
+                }
+
+                await scheduleMessageDelivery(operation, DateTimeOffset.UtcNow + delay.Value, cancellationToken).ConfigureAwait(false);
                 return;
             }
 
@@ -128,6 +137,47 @@
             await Send(wrapper, sendQueue, timeToBeReceived ?? CloudQueueMessageMaxTimeToLive).ConfigureAwait(false);
         }
 
+        internal static TimeSpan? GetVisibilityDelay(List<DeliveryConstraint> constraints)
+        {
+            var doNotDeliverBefore = FirstOrDefault<DoNotDeliverBefore>(constraints);
+            if (doNotDeliverBefore != null)
+            {
+                return ToNullIfNegative(doNotDeliverBefore.At - DateTimeOffset.UtcNow);
+            }
+
+            var delay = FirstOrDefault<DelayDeliveryWith>(constraints);
+            if (delay != null)
+            {
+                return ToNullIfNegative(delay.Delay);
+            }
+
+            return null;
+        }
+
+        static TDeliveryConstraint FirstOrDefault<TDeliveryConstraint>(List<DeliveryConstraint> constraints)
+            where TDeliveryConstraint : DeliveryConstraint
+        {
+            if (constraints == null || constraints.Count == 0)
+            {
+                return null;
+            }
+
+            for (var i = 0; i < constraints.Count; i++)
+            {
+                if (constraints[i] is TDeliveryConstraint c)
+                {
+                    return c;
+                }
+            }
+
+            return null;
+        }
+
+        static TimeSpan? ToNullIfNegative(TimeSpan value)
+        {
+            return value <= TimeSpan.Zero ? (TimeSpan?)null : value;
+        }
+
         Task Send(MessageWrapper wrapper, QueueClient sendQueue, TimeSpan timeToBeReceived)
         {
             string base64String;
@@ -173,7 +223,8 @@
         }
 
         readonly MessageWrapperSerializer serializer;
-        readonly Func<UnicastTransportOperation, CancellationToken, Task<bool>> shouldSend;
+
+        readonly Func<UnicastTransportOperation, DateTimeOffset, CancellationToken, Task> scheduleMessageDelivery;
 
         readonly QueueAddressGenerator addressGenerator;
         readonly AzureStorageAddressingSettings addressing;
